Adapt the plugin to quoted text escaping new behavior of QOwnNotes (#80)
* New method to maintain compatibility with auto tagging: transforming back the @starx/@endx

Auto tagging plugin transforms @anyword to tags, for the @startwhatever and @endwhatever in plantuml we transform them back to the original text.
It allows compatibility with Auto tagging
It allows use of others start/end keywords than startuml/enduml (such as @startgantt)

Also a small change to allow "'" to be used in Gantt.

* Bump version

* Adapt the plugin to quoted text escaping new behavior of QOwnNotes

QOwnNotes has a new behavior for quoted text: all special characters are escaped in numerical entities (cf. https://github.com/pbek/QOwnNotes/issues/1643). This is not compatible with PlantUML.

I made some changes to unescape text and preserve \n

* Bump version number