Adds context menu for inserting head/footer of diagrams (#159)
* Closer to the official Critic Markup spec

We are still not exactly as the spec for 2 reasons:

1. I introduced (previous commit) a multi-line markup possibility that is not in the spec but that is quite handy
2. I am still struggling with a too greedy regex (L. 71) to deal with {~~something~>\n\n~~}

* Bump version number

* Update description and version number

* Remove Node useless dep + introduce async call PUML

Still buggy because regenerateNotePreview seems not able to refresh images (a cache somwhere?)

* Introducing SVG output option

* If workDir option is undefined set a proper cache

* oops!

* Correct image ext depending the SVG Option

* Introducing a rudimentary cache system

No need to recalculate all diagrams for every hits. Diagrams are recalculated only if the md5 hash of the old ones are the same as the new ones.

* Using flags to avoid endless loop with regenerateNotePreview

Should be working now... Needs heavy testing.

* better handling of the flag

should work now?

* \n not properly unescaped

* Unescape parenthesis

* Introducing timestamped callbacks and tidying

Ready for PR?

* Trying to solve https://github.com/qownnotes/scripts/issues/123

By using toNativeDirSeparators

If you use Windows, please test!

* bump new vesrion

* Adds context menu for inserting head/footer of diagrams

* bump version number