Added features to export-notes-as-website and add some dev notes (#182)
* Added export options to export-notes-as-website.

This version has the option of using the old behavior (export inside notebook folder) but with the added ability to customize the export folder name. It also adds a selection to export to a centrally decided folder, where each notebook exports in a subfolder named after that notebook.

* Update info.json

Adjusted the version number and the minimum application version to reflect the use of the new "directory" setting.

* Update README.md

Adding how to edit scripts in the github web interface.

* Update export-notes-as-website.qml

Fixed the update for default behaviors if defaultExportFolder or defaultExportPath are empty.