ollama-integration: rename to ai-local-backend-integration
Any OpenAI-compatible backend is supported, so it would be better to
maintain them in a single script and have seperate scripts for the ones
that aren't compatible.

This would also be clearer for users which might wrongly assume that
only one backend is supported.

This change also automatically adds the OpenAI endpoint to the base URL
because all compatible backends support it already. Therefore, users
only need to provide the local address.